wrong variable size

+++ src/arch/arm-lpc2xxx/interrupt.c
@@ -41,7 +41,7 @@
 
 void interrupt_handler(void);
 
-static uint_fast8_t lpc_family = 0;
+static uint32_t lpc_family = 0;
 #define LPC2229_ID  33685267
 #define LPC2292_ID  67239699
 #define LPC2364A_ID 100924162
